# MyVerse.API/MyVerse/bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "MyVerse",
    "version": "v1"
  },
  "paths": {
    "/api/Security": {
      "post": {
        "tags": [
          "Security"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "user",
            "type": "string"
          },
          {
            "in": "query",
            "name": "password",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Verses/Get": {
      "get": {
        "tags": [
          "Verses"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Verses/Post": {
      "post": {
        "tags": [
          "Verses"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Verses"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Verses/Put": {
      "put": {
        "tags": [
          "Verses"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Verses"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Verses/Delete": {
      "delete": {
        "tags": [
          "Verses"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "Verses": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "book": {
          "type": "string"
        },
        "verse": {
          "format": "int32",
          "type": "integer"
        },
        "chapter": {
          "format": "int32",
          "type": "integer"
        },
        "text": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the bearer scheme.\r\n\r\n Enter 'Bearer' [space] and then yout token in the text input below.\r\n\r\nExample: \"Bearer 12345abcdef\""
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}